fix: vca performance
